docs(devframe): document rpc/ folder + WeakMap context pattern
Skill and rpc guide now describe the one-file-per-RPC layout under
src/rpc/ and the WeakMap-based src/context.ts for sharing setup-time
state across files. All three examples adopt the pattern, with
streaming-chat demonstrating the WeakMap context for its shared
channel + history.
