+++ examples/streaming-chat/src/constants.ts
@@ -1,0 +1,9 @@
+export const CHANNEL_NAME = 'devframe-streaming-chat:tokens'
+export const HISTORY_KEY = 'devframe-streaming-chat:history'
+export const MAX_HISTORY = 200
+
+export const DEMO_PROMPTS = [
+  'Tell me about devframe.',
+  'How does streaming work?',
+  'Write a haiku about RPC.',
+] as const
