+++ examples/streaming-chat/src/rpc/demo-prompts.ts
@@ -1,0 +1,9 @@
+import { defineRpcFunction } from 'devframe'
+import { DEMO_PROMPTS } from '../constants'
+
+export const demoPrompts = defineRpcFunction({
+  name: 'devframe-streaming-chat:demo-prompts',
+  type: 'static',
+  jsonSerializable: true,
+  handler: () => ({ prompts: [...DEMO_PROMPTS] }),
+})
